Removed reference to the removed mpl backend
Noticed in the [release notes](http://vispy.org/releasenotes.html) that the `mpl_plot` module has been removed, but there was still a reference to it in the readme.  I have not made sure that the rest of the modules are still there, or if other modules have been added.